requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
asyncpg>=0.29.0
openai>=1.30.5
pydantic>=2.7.2
python-dotenv>=1.0.1
supabase>=2.5.0
